+++ addon-test-support/only.js
@@ -1,0 +1,9 @@
+import { only as emberQUnitOnly } from 'ember-qunit';
+import { wrapTest, commonConfig } from './utils/config';
+
+// Global sinon config setup
+commonConfig();
+
+export default function only(testName, callback) {
+  wrapTest(testName, callback, emberQUnitOnly);
+}
